requirements.txt
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
